#22 Clang warnings elimination & Conversion from old C-style casts to C++ casts

+++ main.cpp
@@ -39,11 +39,11 @@ int main(int argc, char *argv[])
     Parameters.LoadConfig(app, argc, argv);
 
     //Creation of the two most important objects:
-    MathMod mm(nullptr, Parameters.MaxTri, Parameters.MaxPt, Parameters.MaxGrid,
-               Parameters.NbComponent, Parameters.NbVariables, Parameters.NbConstantes,
-               Parameters.NbDefinedFunctions, Parameters.NbTextures, Parameters.NbSliders,
-               Parameters.NbSliderValues, Parameters.Threads[0], 40,
-               Parameters.CalculFactor[0], Parameters.CalculFactor[1], Parameters.CalculFactor[2]
+    MathMod mm(nullptr, uint(Parameters.MaxTri), uint(Parameters.MaxPt), uint(Parameters.MaxGrid),
+               uint(Parameters.NbComponent), uint(Parameters.NbVariables), uint(Parameters.NbConstantes),
+               uint(Parameters.NbDefinedFunctions), uint(Parameters.NbTextures), Parameters.NbSliders,
+               Parameters.NbSliderValues, uint(Parameters.Threads[0]), 40,
+               uint(Parameters.CalculFactor[0]), uint(Parameters.CalculFactor[1]), uint(Parameters.CalculFactor[2])
             );
 
     DrawingOptions drawingopt;
